Split up Submap component
